Lab4/requirements.txt
uvicorn
fastapi
httpx
pydantic
hazelcast-python-client
